Updating dependencies to require sfn 3.0
At some point in the 2.x releases it seems that our template callback
became broken, causing template validation failures and `--print-only`
output that does not reflect the final scrubbed template.

Pinning our dependencies to >= 3.0, < 4.0 should resolve this issue,
but is also a breaking change which merits a major version number change
in the next release.